logreader: make LogReader struct public
